backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "numpy>=2.3.0",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.10.5",
    "pydantic-settings>=2.9.1",
    "pytz>=2025.2",
    "requests>=2.32.4",
    "sentence-transformers>=4.1.0",
]
